add reusable a3 import command mixin
